Fix handling of remote error on a ghost future
Do not re-rise the exception received, just log it. Re-rising breaks
the protocol of `set_exception` function, which normally would not
rise the exception itself, but rate make it available to whom is
waiting on the future. Rising exception from `_handle_error` breaks
the whole processing loop for a connected client.